Accessibility: Speech for all toolbar buttons
Give every toolbar button an accessible name and description for use
with screen readers. By default, the name and description matches the
tooltip title and description, but they can be different if required
for a specific button.

Tooltips for accidentals in the Note Input toolbar have been updated
to use words (e.g. 'sharp', 'flat', etc.) instead of the relevant
Unicode symbol (♯, ♭). This is necessary to enable screen readers to
say the name correctly. It is also an improvement for sighted users
who might not know the correct name for the symbol. The symbol itself
is displayed as the button icon in the toolbar so there is no need to
also show it in the tooltip.

https://github.com/musescore/MuseScore/pull/8942