Shortcuts: Use F2 for editing and Shift+F2 for Special Chars dialog
Use F2 as an alternative to Alt+Shift+E to enter edit mode to edit text
and other elements. F2 is the standard shortcut to edit cell data in
spreadsheet applications and to edit file names in File Explorer.

MuseScore 3 used F2 to open the Insert Special Characters dialog, so
that is changed to use Shift+F2 instead.

https://github.com/musescore/MuseScore/pull/11576