Implement ButtonBox UI Component
This commit implements a UI Component- ButtonBox. You use ButtonBox to provide the buttons in the bottom panel of dialogs

I have tried my best to replicate Qt's QDialogButtonBox

Usually bottom panels contain buttons of similar style. As a result ButtonBox manages these buttons for you

An added benefit is that ButtonBox will arrange the buttons to be inserted in an order that is consistent with the user's operating system

https://github.com/musescore/MuseScore/pull/19429