Workaround for triggering shortcuts for non-English keyboard layouts
For some unknown reason, QCocoaKeyMapper::possibleKeys does not work correctly for non-English keyboard layouts.
For example, if you press Ctrl+Z on the Russian layout, then the result of the possibleKeys method will be 'Ctrl+Я' and 'Z'. Because of this, qt cannot find shortcuts for these combinations.

The workaround is to register a shortcut depending on the current keyboard layout for possible keys - 'Ctrl+Я' and 'Ctrl+Z'

https://github.com/musescore/MuseScore/pull/10857