Add a switch to disable MIDI 2.0 output manually
A use case for this is the following: on macOS, the MIDI IAC Driver reports that it supports MIDI 2.0, but if you connect any MIDI 1.0 apps to that driver, those apps won't understand the MIDI 2.0 messages. In that case, you want to force MuseScore to produce MIDI 1.0 output even though the receiver reports that it it supports MIDI 2.0.

https://github.com/musescore/MuseScore/pull/12329