Fix re-translatability for theme names
Theme names are initialized once and then stored forever as part of the theme values. This means that after changing the ui language, the theme names won't be re-translated even after restarting MuseScore. Instead, determine the theme name based on the theme code, in a way that is ready to handle user-defined themes in the future.

https://github.com/musescore/MuseScore/pull/13743