Show tab bar instead of title bar for horizontal panels
KDDockWidgets requires using two MouseAreas: one for dragging the title bar (which will drag the whole floating window when it's undocked) and one for dragging individual tabs (this will drag the tabs out of the floating window, which gives a strange effect when there is only one tab in the floating window.

The design specifies that we want tabs instead of a title bar for horizontal panels even when there is only one tab. In this case, we need to make the tabs visible, but make their MouseArea disabled, and instead use the title bar MouseArea.

https://github.com/musescore/MuseScore/pull/22050