Include sndfile include directory privately
When using HomeBrew sndfile, the include directory for libsndfile is `/opt/homebrew/include` or `/usr/local/include`. This folder contains a lot of other stuff (for example, also the HomeBrew version of Qt) that may conflict with our own headers (or the Qt headers of the version that is actually being used).

https://github.com/musescore/MuseScore/pull/24674