Update PCH - use canonical names, trim a bit, add missing
When some std header is "internally" included by one of our headers, and that header of ours is included a lot of times, it may be beneficial to pre-compile that std header. But it won't appear in iwyustat output: that only says that the header of ours is included 998 times, but not that that means that we should precompile that std header.

(IWYU is simply not the right tool to use, actually. What we need to know is: in how many *translation units* is this header used. What IWYU tells us: in how many *files* is this header used. Sometimes this leads to massive overestimations, for example when the same header should be included in all headers of a certain include chain, but that means that it's still only used in one translation unit. Sometimes it leads to massive underestimations, as explained above (and therefore I added some headers manually).)

https://github.com/musescore/MuseScore/pull/25146