set `CMAKE_MSVC_RUNTIME_LIBRARY` to "MultiThreadedDLL"
to prevent build failures and/or crashes in debug builds because debug/release mismatches: for Qt, we always use release builds, and therefore we'll also have to build against the release MSVC runtime, even in debug mode. This means we're missing out on useful debug assertions, but it's the only way for now.

https://github.com/musescore/MuseScore/pull/25549