Use default dock visibility
There might be docks whose visibility is controlled by the page model - meaning they are visible by default, but the page can override and hide them. As a result, there’s a mismatch in visibility, which matters for the dock system when building the page. The actual visibility will be restored later when the page state is restored.

https://github.com/musescore/MuseScore/pull/28228