iex/midi: Improve instrument detection heuristic
The instrument detection heuristic relied on the old order of the
instruments.xml (changed by 28fa244). This resulted in (slightly)
different instruments to be preferred (5 string bass instead of regular etc.).
This commit introduces preferred instruments for each
General MIDI Level 1 instrument to solve that.

Percussion was affected more by the instrument order change. Here, it resulted
in completely wrong instruments to be selected by findClosestInstrument.
(Automobile brake drums or Bongos instead of drumsets).
This is solved by changing the findClosestInstrument algorithms for percussion
to prefer instruments that can play all percussion notes.

https://github.com/musescore/MuseScore/pull/28487