FlatRadioButton: don't use `contentItem` for custom content components
From https://doc.qt.io/qt-6/qml-qtquick-controls-control.html#contentItem-prop :

> The content item is automatically positioned and resized to fit within the padding of the control. Bindings to the x, y, width, and height properties of the contentItem are not respected.

That is what caused https://github.com/musescore/MuseScore/issues/28450; the Column containing the icon and text was supposed to be centered, but was instead stretched to fill the FlatRadioButton.

As a solution, we don't use `contentItem` anymore in those cases, and just place the custom component freely as a child of the `FlatRadioButton`.

Resolves: https://github.com/musescore/MuseScore/issues/28450

https://github.com/musescore/MuseScore/pull/28812