fixed #28753: KDDW patch. Ensures stack layout is polished after insertion
This commit added invalidate for QQuickStackLayout during creation and removed stack rearrangement when new elements are added:
https://github.com/qt/qtdeclarative/commit/4de9df2c9ae0ba35ad425d46fc9f2e982e0da404

When restoring the frame geometry, we also set the size for the StackLayout, but because it’s invalidated, rearrangement for the stack is ignored.
The solution is to restore the previous behavior — after adding an item, update the StackLayout.

https://github.com/musescore/MuseScore/pull/28832