Don't restore window as minimised
Resolves: https://github.com/musescore/MuseScore/issues/30388

Unminimising a window turns out to be difficult, if you want that the geometry is preserved when the window filled the whole screen before being minimised. Qt considers a window that happens to fill the whole screen as “maximised”; it seems that when unminimising a window, either using `show` or `setVisibility(AutomaticVisibility)`, Qt wants it to return to a “normal” state, which is not a maximised state, so the window is made a bit smaller (and it’s ugly). 
A plain call to `window->setWindowStates(states & ~Minimised)` seems to do exactly nothing.

Instead of trying to unminimise the window, we just make sure it never gets into a minimised state at all, during geometry restoration.

https://github.com/musescore/MuseScore/pull/30400