Fix and improve build_without_qt
In the end, the problem was that iex_musicxml was not linked with muse_global, because that only happened when `MODULE_LINK_GLOBAL` is `ON`, and that is coincidentally the case in the default build but not in `build_without_qt`. This commit removes `MODULE_LINK_GLOBAL` entirely, and cleans up the `build_without_qt` CMakeLists.txt a little bit.

https://github.com/musescore/MuseScore/pull/31580